[minor] Split iceberg table manager
