+++ src/moonlink/src/storage/iceberg/mod.rs
@@ -1,6 +1,8 @@
 pub(super) mod deletion_vector;
 pub(super) mod file_catalog;
+pub(super) mod iceberg_table_loader;
 pub(super) mod iceberg_table_manager;
+pub(super) mod iceberg_table_syncer;
 pub(super) mod index;
 pub(super) mod moonlink_catalog;
 pub(super) mod parquet_metadata_utils;
